adjust max tasks per child to be 200
Need to adjust for concurrency ie right now it would wait until a process hit 6000 tasks. Since the general queues have 30 processes, it would take a while for just one to hit 6000. I adjusted it down to 200; 6000/30. The express queue has 4 processes so it will cycle quicker. If that becomes a problem we can use the cli to adjust for each celery worker.