use cli arg for max tasks per child for celery
It is possible the settings.py setting isn't working for some reason. I want to double check using the cli arg to confirm. It is also possible this setting just isn't addressing our problem here. I didn't touch the express script either so if this works we can add this to that as well.